Add compilation script for Linux, reorganize tools directory.
- Add compile.sh, compile_fdisk.sh and command/msxdos/compile.sh.
  These allow to build Nextor on Linux, using the M80dotNet project
  (https://github.com/Konamiman/M80dotNet) as well as native Linux
  versions of sjasm and mknexrom.
- Reorganize tools directory. Now instead of a "wintools" directory,
  there's a "buildtools" directory with subdirectories for the sources
  and for Windows and Linux binaries.
