+++ .gitignore
@@ -47,3 +47,4 @@ source/tools/data.mac
 *.suo
 NEXTOR.SYS
 NEXTORK.SYS
+nextor_base.dat
