+++ buildtools/windows/SymToEqus.exe
