+++ source/kernel/bank5/.gitignore
@@ -3,3 +3,4 @@ condasm.inc
 const.inc
 macros.inc
 printf.cs
+*.relx
