--- wintools/mksdscc.bat
@@ -1,1 +1,0 @@
-dd if=kernel.dat of=sccsd.drv skip=115456 count=16384 bs=1
